app: move POST /api/config to app.routers.config
